优化xml文件，删除无用参数

+++ app/src/main/java/com/lalilu/lmusic/fragment/AlbumDetailFragment.kt
@@ -51,7 +51,7 @@ class AlbumDetailFragment : DataBindingFragment(), CoroutineScope {
         }
         mAdapter.onItemLongClick = {
             findNavController().navigate(
-                AlbumDetailFragmentDirections.albumToSongDetail(it.mediaId.toLong())
+                AlbumDetailFragmentDirections.albumToSongDetail(it.mediaId)
             )
         }
         return DataBindingConfig(R.layout.fragment_detail_album)
